feat: add report data source
Introduce a new "armis_reports" data source for retrieving report
information from Armis, with optional filtering by report ID or name.
Includes schema, implementation, and acceptance tests. Also updates
the Schedule model to use float64 for RepeatAmount.

+++ armis/model_reports.go
@@ -35,7 +35,7 @@ type Report struct {
 // Schedule represents a report schedule.
 type Schedule struct {
 	Email            []string `json:"email,omitempty"`
-	RepeatAmount     int      `json:"repeatAmount,omitempty"`
+	RepeatAmount     float64  `json:"repeatAmount,omitempty"`
 	RepeatUnit       string   `json:"repeatUnit,omitempty"`
 	ReportFileFormat string   `json:"reportFileFormat,omitempty"`
 	TimeOfDay        string   `json:"timeOfDay,omitempty"`
